pytests/iroha_cli_tests/pyproject.toml
[tool.poetry]
name = "iroha_cli_tests"
version = "0.1.0"
description = ""
authors = ["alexstroke <busyfifer@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
faker = "*"
allure-python-commons = "*"
cryptography = "*"
python-dotenv = "*"
tomlkit = "^0.13.2"

[tool.poetry.dev-dependencies]
pytest = "*"
allure-pytest = "*"
pytest-xdist = "*"
mypy = "*"
black = "*"
flake8 = "*"

[tool.poetry.group.dev.dependencies]
pylint = "^3.2.7"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"



